Generalized the used Traits for the Matrix Implementation
Instead of a hard dependency on floats a more generous approach for the
implementation is now chosen, which more explicitly enforces "feature"
traits instead of the types it works on. Essentially this does not make
a difference in real flexibility, as we will only be using floats.
However it eases certain implementation restrictions, where code is
based on Matrices.