Removed num_traits::Float from matrix implementation
Eventhough I learned a lot about traits and the typing/generics system
of Rust switching over to the Float type of num_traits in the end
I realized it is not worth the hassle and the bloated generic structures
everywhere. Essentially we require floating points throughout the
application in almost any data structure. Most likely we would actually
only want to use 64 bit precision floating points, as there shouldn't be
much of a performance difference and the little performance gain, that
might be there due to the possibility of having a bigger amount of
values in cache, is quickly negated by the hassle we might introduce by
lacking precission. Nevertheless I decided to simply define the used
floating point type globally. This way we are able to switch to 32bit in
order to check if there might be a performance gain, once we are done
implementing real ray tracing ;).