Fixed clippy error regarding the comparison between floats
rust-clippy is a linter for rust, which catches best-practice
violations, or oversights which are easy pitfalls creating bugs.

I am going to try and run clippy more in the future to keep the project
warning and error free. But first I need to remove the errors and
warning currently reported by clippy.