Implemented Intersection Collection and Body dispatching
All the intersections are now housed within an `Intersections` struct,
which holds the corresponding vector internally and exposed the `Index`
trait as well as functions to determine length and emptiness. This was
done to on one hand provide a proper entrypoint to sort the given
collection data, as well as on the other hand to implement a `hit`
method, which always returns the *first* *non-negative* intersection
within the collection for easier processing.

Furthermore a static dispatching to different kinds of `Body` objects
was implemented, to be used in the future, when there are more Bodies
within the scene than a Sphere. This technique of *enum dispatching*
should be reasonable fast as well as flexible enough for us to introduce
furhter types in the future.