Removed ppm output from chapter05 example
The ppm output is quite slow currently, as it does a lot of copying data
around during the export process. This could of course be optimized, but
for now we have a quite fast png output, which works well for us.
Therefore the ppm output is not really needed at this point. We may
revisit this in the future.