Extract world to object conversion from intersection calculation
The transformation from world space into object space and vice versa is
a necessity for the intersection and normal calculation of every body.
Those which are already there (Sphere) and those, which we are going to
implement in the future. Therefore this functionality was extracted into
the Trait, to be used within every further Body object, still to be
implemented.