Implement FuzzyEq for World and Camera
In order to implement fuzzyeq for arrays the cloning of the given value was needed in order to not take it. Otherwise everything could only have been compared by reference. Eventhough this might be the better solution, I decided for now to use cloning, as fuzzyeq is only used for tests, which aren't performance critical.