Implemented a yaml based world/scene parser
A world with multiple cameras may now be instantiated based on
a dynamically read YAML file, which provides all the necessary
information. An example for such a YAML file can be found within the
tests of the `world_loader::Yaml` struct. It looks something like this:

```yaml
---
- camera:
    name: output1
    width: 800
    height: 600
    field_of_view: 0.785 # 45degrees
    from: [1, 2, 3.4]
    to: [5.6, 7, 8]
    up: [9.10, 11, -1.2]

- light:
    type: point_light
    at: [1.1, 2.2, 3.3]
    intensity: [0.4, 0.5, 0.6]

- body:
    type: sphere
    material:
      type: phong
      color: [1,1,1]
      diffuse: 0.7
      ambient: 0.1
      specular: 0.0
      shininess: 200
    transforms:
      - type: translate
        to: [1, 2, 3]
      - type: rotate_x
        radians: 3.14
```

Currently the parser is not used anywhere besides from its tests.

The future will most likely contain a way to predefine transforms and
materials, to be reused by identifier throughout the world definition.

The parser is currently using a big amount of macros as well as not very
need transport of the parsing state, from invocation to invocation.
I wanted to get the thing going first and later on think about ways to
make it nicer. Pull requests and/or ideas on how to improve this are
very welcome. The rust typing still felt a little bit alien to me, while
I wrote this code.