Implemented renderer executable using the YAML world loader
The new executable can be provided with a world.yaml file, which
contains a whole scene definition including lights, bodies and multiple
cameras. For each defined camera a resulting png will be rendered.

Furthermore a yaml world definition of the 3 spheres scenes we used
before is included in this. A second camera has been added and the
materials have been altered.