Implement basic usage of Pattern in Material (Phong)
The way this is done might change in the near future. Currently the
pattern overwrites the color, if it is present. However I don't really
like that, as a solid color is nothing else, than a static pattern.
Furthermore I don't like the possibility to even specify a normal color
if a pattern is present, which is then not used at all, as this might
produce strange hard to decipher behaviour.