Implement YAML syntax for defining striped pattern
The YAML world loader now supports the loading of patterns within
material contexts. For now the Striped pattern has been implemented, as
it is the only one which has been created yet. The syntax should
essentially be generic enough to allow for the implementation and
parsing of other patterns in the future.