Implement pattern transformation into object space
To implement object space transformation of patterns it is required for
them to know the object (body) transform. In order to pass this
information we are passing along a reference to the hit body object.

Transformation into pattern space via custom transform of a specific
pattern is marked as a fixme.