Implemented calculation of reflection colors with recursion limit
As reflections may be endless (Think about two planes oposite of each
other), we are limiting the maximum amount of reflection recursion.
A value 5 seems to be quite adequate at the moment, and is currently
hardcoded. It should most likely be made configurable later on,
therefore it is stored within the world struct.